pdf-processor-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.8
Pillow==10.1.0
openai==1.3.7
python-dotenv==1.0.0
mangum==0.17.0